refactor(TransactionFinish): code
