+++ src/i18n.ts
@@ -267,6 +267,12 @@ export enum I18N {
   transactionConfirmationSendTo,
   transactionConfirmationTotalAmount,
   transactionConfirmationSum,
+  transactionFinishSendingComplate,
+  transactionFinishNetworkFee,
+  transactionFinishEditContact,
+  transactionFinishAddContact,
+  transactionFinishHash,
+  transactionFinishDone,
   onboardingFinishCreate,
   onboardingFinishRecover,
   walletCreateAddAccount,
@@ -651,6 +657,12 @@ const en: Record<I18N, string> = {
   [I18N.transactionConfirmationSendTo]: 'Send to',
   [I18N.transactionConfirmationTotalAmount]: 'Total Amount',
   [I18N.transactionConfirmationSum]: '{{sum}} ISLM',
+  [I18N.transactionFinishSendingComplate]: 'Sending Completed!',
+  [I18N.transactionFinishNetworkFee]: 'Network Fee: {{fee}} ISLM',
+  [I18N.transactionFinishEditContact]: 'Edit Contact',
+  [I18N.transactionFinishAddContact]: 'Add Contact',
+  [I18N.transactionFinishHash]: 'Hash',
+  [I18N.transactionFinishDone]: 'Done',
   [I18N.settingsAccountNoWallet]: 'No wallets',
   [I18N.LedgerStoreWalletSaving]: 'Account saving in progress',
   [I18N.ledgerVerifyAddress]:
